Remove dupState().
It was used due to a misunderstanding in my part about interfaces+pointers.
I was afraid the user might be able to mess with the internal state by
modifying the returned states. This is the case if they make use of pointers
to states/ops. If they use structs then everything will be copies.
This is expected go behaviour and this gives users the freedom to mess around.

License: MIT
Signed-off-by: Hector Sanjuan <hector@protocol.ai>