Support forced rollbacks. Detect unconsistent states.
This implements a "rollback" operation which simply consists
in hammering the new state with a special log operation which is
very similar to CommitState().

This aditionally introduces handling of decoding/ApplyTo() errors,
by marking the state as inconsistent and failing to return it anymore
with GetLogHead(). Inconsistent states are fixed with a Rollback().

License: MIT
Signed-off-by: Hector Sanjuan <hector@protocol.ai>