Do not export Encode/DecodeState but Encode/DecodeSnapshot
encodeState and decodeState were meant as internal operations for
how we serialize states for transmitting them with raft.
It turns out, they were not used in the end and instead encodeOp and
decodeOp (which are very similar) took their place.

As convinience, we used encodeState and decodeState as usable
serializable format for snapshots. How snapshots are formatted
is something useful to export, so users can work directly with them.
But that format should not be tied to how states are [de]serialized
for raft (an internal choice), which is what encodeState and
decodeState were for in the first place (and might be used
for in the future).

For example, you could use msgpack for sending states and operations,
but simply use json for snapshot formatting. This change allows more
freedom if we are ever to change how things are serialized internally
vs how snapshots are stored.

License: MIT
Signed-off-by: Hector Sanjuan <hector@protocol.ai>