FSM: Eliminate snapshot buffering
Currently we forced Marshalling to and from a byte slice instead of
directly sending the encoded State to the Snapshot sink, and decoding
it directly from the Snapshot reader.

For huge states this was obviously very innefficient, as it will cause
an extra in-memory copy of the state.

We also remove the dependency to deprecated go-multicodec. See:

https://github.com/multiformats/go-multicodec/pull/36

Additionally, allow custom serialization for Op.

And opWrapper is no longer necessary.

And things work as they should now with the current serialization
strategy.

stateWrap and opWrap etc. were necessary before because things did
not deserialize properly onto the given objects and probably because
I did not know what I was doing. Hopefully this will actually still
work when plugged to cluster.

License: MIT
Signed-off-by: Hector Sanjuan <hector@protocol.ai>